+++ navigator_helpers/tasks/prompt_templates/__init__.py
@@ -1,1 +1,3 @@
-from navigator_helpers.tasks.prompt_templates.suite import load_prompt_template_suite
+from navigator_helpers.tasks.prompt_templates.template_suite import (
+    load_prompt_template_suite,
+)
