+++ navigator_helpers/tasks/text_to_code/task_suite.py
@@ -1,19 +1,21 @@
 from __future__ import annotations
 
 import logging
-import random
 import re
+import uuid
 
 from enum import Enum
+from typing import Optional
 
 import numpy as np
 
-from pydantic import BaseModel
+from tqdm import tqdm
 
 from navigator_helpers.content_validator import ContentValidator
 from navigator_helpers.logs import get_logger, SIMPLE_LOG_FORMAT
 from navigator_helpers.tasks.prompt_templates import load_prompt_template_suite
 from navigator_helpers.tasks.text_to_code import utils
+from navigator_helpers.tasks.text_to_code.contextual_tags import ContextualTags
 from navigator_helpers.tasks.text_to_code.llm_suite import GretelLLMSuite, LLMSuiteType
 
 logger = get_logger(__name__, fmt=SIMPLE_LOG_FORMAT)
@@ -21,6 +23,14 @@ validator = ContentValidator()
 
 fluff_logger = logging.getLogger("sqlfluff")
 fluff_logger.setLevel(logging.ERROR)
+
+PBAR_TEMPLATE = "Running Pipeline [current task: {}]".format
+
+
+def _update_pbar_desc(pbar: tqdm, desc: str):
+    if pbar is not None:
+        pbar.set_description(desc)
+        pbar.refresh()
 
 
 class CodeLang(str, Enum):
@@ -30,28 +40,6 @@ class CodeLang(str, Enum):
     @property
     def title(self) -> str:
         return self.value.upper() if self == CodeLang.SQL else self.value.capitalize()
-
-
-class ContextualTags(BaseModel):
-    domain_and_topics: dict[str, list[str]]
-    complexity_levels: list[str]
-
-    @property
-    def domains(self) -> list[str]:
-        return list(self.domain_and_topics.keys())
-
-    @property
-    def num_domains(self) -> int:
-        return len(self.domain_and_topics)
-
-    def sample(self) -> tuple[str, str, str]:
-        domain = random.choice(list(self.domain_and_topics.keys()))
-        topic = random.choice(self.domain_and_topics[domain])
-        complexity = random.choice(self.complexity_levels)
-        return domain, topic, complexity
-
-    def __repr__(self):
-        return f"{self.__class__.__name__}({self.model_dump_json(indent=4)})"
 
 
 class NL2CodeTaskSuite:
@@ -261,7 +249,7 @@ class NL2CodeTaskSuite:
         response = utils.parse_json_str(self.llm.judge_generate(prompt)) or {}
         for k, v in response.items():
             scores[f"{k}_score"] = v["score"]
-            scores[f"{k}_reasoning"] = v["reasoning"]
+            scores[f"{k}_reason"] = v["reasoning"]
         return scores
 
     def eval_sql_with_llm_as_judge(
@@ -277,5 +265,121 @@ class NL2CodeTaskSuite:
         response = utils.parse_json_str(self.llm.judge_generate(prompt)) or {}
         for k, v in response.items():
             scores[f"{k}_score"] = v["score"]
-            scores[f"{k}_reasoning"] = v["reasoning"]
+            scores[f"{k}_reason"] = v["reasoning"]
         return scores
+
+    def create_nl2python_record(
+        self,
+        domain: str,
+        topic: str,
+        complexity: str,
+        llm_as_a_judge: bool,
+        syntax_validation: bool,
+        progress_bar: Optional[tqdm] = None,
+    ) -> dict:
+        _update_pbar_desc(
+            progress_bar, f"⏳ {PBAR_TEMPLATE('suggest python packages')}"
+        )
+        suggested_packages = self.generate_suggested_python_packages(
+            domain, topic, max_dependencies=np.random.randint(5, 8)
+        )
+        _update_pbar_desc(
+            progress_bar, f"⏳ {PBAR_TEMPLATE('natural language prompt')}"
+        )
+        natural_language = self.generate_python_natural_language(
+            domain, topic, complexity
+        )
+        _update_pbar_desc(progress_bar, f"⌛️ {PBAR_TEMPLATE('python code generation')}")
+        prompt, code = self.python_code_generation(
+            natural_language,
+            domain,
+            topic,
+            complexity,
+            ",".join([f" `{dep}`" for dep in suggested_packages]),
+        )
+        record = {
+            "uid": uuid.uuid4().hex,
+            "domain": domain,
+            "topic": topic,
+            "complexity": complexity,
+            "suggested_packages": suggested_packages,
+            "full_prompt": prompt,
+            "natural_language": natural_language,
+            "code": code,
+        }
+        if llm_as_a_judge:
+            _update_pbar_desc(
+                progress_bar, f"⏳ {PBAR_TEMPLATE('llm-as-a-judge evaluation')}"
+            )
+            scores = self.eval_python_with_llm_as_judge(natural_language, code)
+            record.update(scores)
+        if syntax_validation:
+            _update_pbar_desc(progress_bar, f"⌛️ {PBAR_TEMPLATE('syntax validation')}")
+            syntax_validation = self.validate_code(code)
+            record["syntax_validation"] = syntax_validation
+        return record
+
+    def create_nl2sql_record(
+        self,
+        domain: str,
+        topic: str,
+        complexity: str,
+        llm_as_a_judge: bool,
+        syntax_validation: bool,
+        progress_bar: Optional[tqdm] = None,
+    ) -> dict:
+        _update_pbar_desc(progress_bar, f"⏳ {PBAR_TEMPLATE('sql tables and views')}")
+        sql_context = self.generate_sql_tables_and_views(
+            domain, topic, max_statements=np.random.randint(3, 6)
+        )
+        _update_pbar_desc(
+            progress_bar, f"⏳ {PBAR_TEMPLATE('natural language prompt')}"
+        )
+        natural_language = self.generate_sql_natural_language(
+            domain, topic, complexity, sql_context
+        )
+        _update_pbar_desc(progress_bar, f"⌛️ {PBAR_TEMPLATE('sql generation')}")
+        prompt, code = self.sql_code_generation(
+            natural_language, domain, topic, complexity, sql_context
+        )
+        record = {
+            "uid": uuid.uuid4().hex,
+            "domain": domain,
+            "topic": topic,
+            "complexity": complexity,
+            "sql_context": sql_context,
+            "full_prompt": prompt,
+            "natural_language": natural_language,
+            "code": code,
+        }
+        if llm_as_a_judge:
+            _update_pbar_desc(
+                progress_bar, f"⏳ {PBAR_TEMPLATE('llm-as-a-judge evaluation')}"
+            )
+            scores = self.eval_sql_with_llm_as_judge(
+                natural_language, code, sql_context
+            )
+            record.update(scores)
+        if syntax_validation:
+            _update_pbar_desc(progress_bar, f"⌛️ {PBAR_TEMPLATE('syntax validation')}")
+            syntax_validation = self.validate_code(code)
+            record["syntax_validation"] = syntax_validation
+        return record
+
+    def create_record(
+        self,
+        domain: str,
+        topic: str,
+        complexity: str,
+        llm_as_a_judge: bool,
+        syntax_validation: bool,
+        progress_bar: Optional[tqdm] = None,
+    ) -> dict:
+        return getattr(self, f"create_nl2{self.code_lang.value}_record")(
+            domain=domain,
+            topic=topic,
+            complexity=complexity,
+            llm_as_a_judge=llm_as_a_judge,
+            syntax_validation=syntax_validation,
+            progress_bar=progress_bar,
+        )
